enable set filters on tds result panel
