+++ .changeset/three-bananas-wash.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-query-builder': patch
+---
+
+Enable set filters in tds result panel.
